socket を どのページでも繋いでも friends などを受け取れるように login page で socket 繋ぐと socket on の処理がなく、データがセットされなかったので socket on の処理を移動
